pyproject.toml
[project]
name = "legal-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.10",
    "litellm>=1.81.16",
    "numpy>=2.4.2",
    "plotly>=6.5.2",
    "pypdf>=6.7.4",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.3",
    "tavily>=1.1.0",
    "tqdm>=4.67.3",
]
